[CWS] use fake hostname in functional tests
